+++ src/hash_set_sequential.h
@@ -31,7 +31,7 @@ class HashSetSequential : public HashSetBase<T> {
   bool Remove(T elem) final {
     bucket_t<T>& bucket = GetBucket(elem);
 
-    for (int i = 0; i < bucket.size(); i++) {
+    for (size_t i = 0; i < bucket.size(); i++) {
       if (bucket[i] == elem) {
         bucket.erase(i);
         return true;
